+++ tests/numeric/test_LogTransformer.py
@@ -26,7 +26,7 @@ class TestInit(object):
 
         with pytest.raises(
             ValueError,
-            match=re.escape("base should be numeric or None"),
+            match=re.escape("LogTransformer: base should be numeric or None"),
         ):
 
             LogTransformer(
@@ -40,7 +40,7 @@ class TestInit(object):
 
         with pytest.raises(
             ValueError,
-            match=re.escape("base should be strictly positive"),
+            match=re.escape("LogTransformer: base should be strictly positive"),
         ):
 
             LogTransformer(
@@ -211,7 +211,8 @@ class TestTransform(object):
         x = LogTransformer(columns=["a", "b", "c"])
 
         with pytest.raises(
-            TypeError, match=r"The following columns are not numeric in X; \['b', 'c'\]"
+            TypeError,
+            match=r"LogTransformer: The following columns are not numeric in X; \['b', 'c'\]",
         ):
 
             x.transform(df)
@@ -367,7 +368,7 @@ class TestTransform(object):
 
         with pytest.raises(
             ValueError,
-            match=f"values less than or equal to 0 in columns{extra_exception_text}, make greater than 0 before using transform",
+            match=f"LogTransformer: values less than or equal to 0 in columns{extra_exception_text}, make greater than 0 before using transform",
         ):
 
             x.transform(df)