+++ tubular/imputers.py
@@ -69,7 +69,9 @@ class ArbitraryImputer(BaseImputer):
 
         if columns is None:
 
-            raise ValueError("columns must be specified in init for ArbitraryImputer")
+            raise ValueError(
+                f"{self.classname()}: columns must be specified in init for ArbitraryImputer"
+            )
 
         super().__init__(columns=columns, **kwargs)
 
@@ -80,7 +82,7 @@ class ArbitraryImputer(BaseImputer):
         ):
 
             raise ValueError(
-                "impute_value should be a single value (int, float or str)"
+                f"{self.classname()}: impute_value should be a single value (int, float or str)"
             )
 
         self.impute_values_ = {}
@@ -402,7 +404,7 @@ class NearestMeanResponseImputer(BaseImputer):
 
         if n_nulls > 0:
 
-            raise ValueError(f"y has {n_nulls} null values")
+            raise ValueError(f"{self.classname()}: y has {n_nulls} null values")
 
         self.impute_values_ = {}
 
@@ -416,7 +418,7 @@ class NearestMeanResponseImputer(BaseImputer):
             if c_nulls.sum() == 0:
 
                 raise ValueError(
-                    f"Column {c} has no missing values, cannot use this transformer."
+                    f"{self.classname()}: Column {c} has no missing values, cannot use this transformer."
                 )
 
             else: