+++ CHANGELOG.rst
@@ -23,6 +23,8 @@ Added
 ^^^^^
 - added support for weights to mean, median and mode imputers `#47 <https://github.com/lvgig/tubular/pull/47>`_
 
+- added classname() method to BaseTransformer and prefixed all errors with classname call for easier debugging `#48 <https://github.com/lvgig/tubular/pull/48>`_
+
 Fixed
 ^^^^^
 - updated black version to 22.3.0 and flake8 version to 5.0.4 to fix compatibility issues `#45 <https://github.com/lvgig/tubular/pull/45>`_