+++ CHANGELOG.rst
@@ -16,7 +16,7 @@ Subsections for each version can be one of the following;
 
 Each individual change should have a link to the pull request after the description of the change.
 
-0.3.3 (Unreleased)
+0.3.3 (2023-01-19)
 ------------------
 
 Added